modified:   Controller/DataLoadController.py 修复启动异常 	modified:   Controller/DataShowController.py 修复启动异常 	modified:   Model/Data/TypeDataList.py 修复启动异常 	modified:   Model/Dataset/DataUnit.py 修复启动异常 	modified:   Model/Dataset/Label.py 修复启动异常 	modified:   Service/DataShow.py 修复启动异常 	modified:   Service/DataTag.py 修复启动异常 	modified:   View/GraphWidget.py 修复启动异常 	modified:   View/MainUI.py 修复启动异常 	modified:   View/MainUI.ui 修复启动异常 	modified:   View/view.py 修复启动异常

+++ Model/Dataset/Label.py
@@ -11,7 +11,7 @@ class Label:
         @param actionType 动作类型
         @param actionEval 动作评价
     """
-    def __init__(self, positionX: int, positionY: int, actionType: str, actionEval: str) -> None:
+    def __init__(self, positionX: float, positionY: float, actionType: str, actionEval: str) -> None:
         self.positionX = positionX
         self.positionY = positionY
         self.actionType = actionType
